sweetalert

+++ fornt-end/ptgigamega/src/components/tableHome.vue
@@ -20,7 +20,25 @@ props:["book","index"],
 methods:{
     ... mapActions(useCounterStore,["deleteBook","getDataById"]),
     removeBook(id){
-        this.deleteBook(id)
+        this.$swal({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+        if (result.isConfirmed) {
+            this.deleteBook(id)
+        this.$swal(
+            'Deleted!',
+            'Your file has been deleted.',
+            'success'
+        )
+        }
+        })
+        
     },
     editBook(id){
         this.getDataById(id)
